listo

+++ database/migrations/2023_04_04_212948_create_abonos_table.php
@@ -16,7 +16,7 @@ class CreateAbonosTable extends Migration
         Schema::create('abonos', function (Blueprint $table) {
             $table->id();
             $table->foreignId('prestamo_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
-            $table->float('monto_abono')->unsigned()->default(0);
+            $table->decimal('monto_abono')->unsigned()->default(0);
             $table->date('fecha')->nullable();
             $table->string('lat', 100)->nullable();
             $table->string('long', 100)->nullable();
